Update Manuscript_COD_Trends.Rmd
Hey Corinne-
I have not made a full pass through this yet, but tried to just give a couple of comments on the Intro and Results. Generally speaking, I think this the right way to go in terms of describing the results--report on the general pattern among all states but identify some key outliers of interest, particularly if they are regional in nature, which could be linked to other changes in risk factors, etc.